Code Restructure: Lexical Scanner + Utilities Module
