+++ olox/Olox.java
@@ -1,5 +1,9 @@
 package olox;
 
+import lexical_scanner.Scanner;
+import lexical_scanner.Token;
+import utils.ErrorReporter;
+
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
@@ -8,11 +12,11 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
 
-import static olox.ExitCode.COMMAND_LINE_USAGE_ERROR;
-import static olox.ExitCode.USER_DATA_INCORRECT;
+import static utils.ExitCode.COMMAND_LINE_USAGE_ERROR;
+import static utils.ExitCode.USER_DATA_INCORRECT;
 
 public class Olox {
-    static ErrorReporter errorReporter = new ErrorReporter();
+    static ErrorReporter errorReporter = ErrorReporter.getInstance();
     public static void main(String[] args) throws IOException {
         if(args.length >  1) {
             System.out.println("Usage: jolox [script_name.lx]");
@@ -28,7 +32,7 @@ public class Olox {
     private static void runFile(String path) throws IOException {
         byte[] bytes = Files.readAllBytes(Path.of(path));
         run(new String(bytes, Charset.defaultCharset()));
-        if(errorReporter.hadError) System.exit(USER_DATA_INCORRECT.code);
+        if(errorReporter.getErrorStatus()) System.exit(USER_DATA_INCORRECT.code);
     }
 
     private static void runPrompt() throws IOException {
@@ -40,15 +44,13 @@ public class Olox {
             String line = reader.readLine();
             if (line == null) break;    // CTRL-D exits interactive loop (CMD-D for Mac)
             run(line);
-            errorReporter.hadError = false;    // Ensure error flag is reset for each interactive command
+            errorReporter.setErrorStatus(false);    // Ensure error flag is reset for each interactive command
         }
     }
 
     private static void run(String source) {
         Scanner scanner = new Scanner(source);
-
         List<Token> tokens = scanner.scanTokens();
-
         tokens.forEach(System.out::println);
     }
 }
